refactor: consolidate all PID logic into bin/lib/pid-manager.js
Single source of truth for all process/PID operations:
- writePid, readLock, removeLock (lock file CRUD)
- isAlive, killPid, killPortOccupant (process management)
- fetchHealth (proxy health check)
- registerCleanup (gateway exit handler)

Both gateway (anthropic-gateway.ts) and launcher (proxy-launcher.js)
import from the same pid-manager.js. No more duplicated code.

Deleted adapters/pid-manager.ts — replaced by bin/lib/pid-manager.js.
